Introduce a generalized query interface
This interface unifies the existing interfaces through a couple of traits.
Also, it uses rusts type system to tie paths to the respective query.
Previously, paths were just constructed from whatever the current state was.
Now, they belong to a result.
This also forces that paths can only be constrcuted, if there actually is a path.